Refactor ConstantLoadInsn class
